settings: Change USER_AGENT into a PC user agent
Change `USER_AGENT` in settings.py, because when we still
use the prior `USER_AGENT` which is an Android user agent, Facebook
enforces most of the spiders to use the mobile version (m.facebook.com)
by redirecting it.

So we change it to use PC user agent instead, in order to be
able to access both version of Facebook (PC and mobile).